requirements-rag.txt
# RAG Layer Dependencies
langchain>=1.0.0
langchain-community>=0.3.0
langchain-core>=1.0.0
langchain-huggingface>=1.0.0
langchain-text-splitters>=1.0.0
sentence-transformers>=5.0.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
PyYAML>=6.0
Jinja2>=3.0